router and hash
hooked up router and hash modules. hash watches for changes to the hash url, and dispatches a message when it does.
router waits for messages from the hash dispatcher, and sets the state of the application based on it.
added a controller to the view and model of resource, to respond to router messages, and capture data, initiliaze the model, and the view