resource edit
wrote and hooked up the edit resource view. sets the state of the current view, then pushes out the state to the hash parser, which takes the new hash and pushes it to the router, which tells the system to redraw.
not the most efficient, since even these state changes that occur within the same view require reaching into the data structure to create a new model, then create a new view from it