resources, tags, classes
resources now reference tags as IDs, and looks to the datastore for their name. perhaps they should be IDed based on the lowercase string representation?
classes now have a router and hash hookup. heading down the path of allowing a user to add a resource to a class. which happens when you are looking at a resource. this will also rely on authentication, which hasn't been sorted out, but that should come soon. at least so kind of temporary identity that gets assigned. doesn't have to actually talk to a server. but can capture what the client side will look like