tags educators classes
made a proper model for tags, which will maintain which resources are referencing it. they are now stored with an ID that comes from the tag name itself. this way its easy to check if a tag already exists, as we want them to be unique
educators now has a model too. and is being referenced in the resource view. these are indexed by their email address, since emails will be unique to each educator who signs up. educators was previously referred to as authors, and prior to that, users.
classes got an update, and now have a presence in the hash-router, so going to a class url is tied up, but still has no real view or viewcontroller