incrementing metadata
localstorage stores strings, not objects. so i have a hacky way of incrementing the count of an object type, by inspecting to see if there is already one in the database. I was under the impression that localStorage had the same implementation as LevelDB, turns out, it does not. Going to fork this into using levelDB