class add controller
working on the class add controller highlights a shortcoming of the datastore gathering function. the need to get data that might not be related, yet, or ever. being able to defer different operations then starting them, and notifying the view when they are all ready is what needs to happen