queue improvements
factored out the bit that sets up the related key dispatching when each model emits its own, and removes itself from the queue. this way if you are only looking for a series of model data that isn't necessarily related to any particular model, you can do so. as is the case with the me model, which is stashed in localstorage, and is a synchronous api, so it doesn't need any callback magic.